+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CLC2v_w2.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CwxUBTKf.js";import{_ as a}from"./mermaid.core-BDS1p4Le.js";import"./chunk-5VM5RSS4-CGOj7d-I.js";import"./chunk-XXDRQBXY-Chl3V-52.js";import"./chunk-POPQ4Y6H-3SEwFus_.js";import"./chunk-F27PBJKO-DU94VBv2.js";import"./channel-CAUKdu_j.js";import"./index-DJrKOLxK.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};