+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-GURpGAb-.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-DaqjvVoy.js";import{_ as a}from"./mermaid.core-DtDprwEV.js";import"./chunk-5VM5RSS4-B3-AoRuh.js";import"./chunk-XXDRQBXY-mUDiU3Sf.js";import"./chunk-POPQ4Y6H-D7IIijWP.js";import"./chunk-F27PBJKO-Cj7wfwzY.js";import"./channel-BtEbTsbH.js";import"./index-xPtTq8-c.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};